+++ lua/CompMod/Marine/AdvancedWeapons/Post/Armory_Server.lua
@@ -1,0 +1,9 @@
+function Armory:OnResearchComplete(researchId)
+
+    if researchId == kTechId.AdvancedArmoryUpgrade then
+
+        self:SetTechId(kTechId.AdvancedArmory)
+
+    end
+
+end