+++ lua/CompMod/Marine/AdvancedWeapons/Post/Balance.lua
@@ -1,0 +1,10 @@
+-- Munitions includes shotgun and heavy machine gun
+kMunitionsTechResearchCost = 25
+kMunitionsTechResearchTime = 45
+
+-- Demolitions includes flamethrower and grenade launcher
+kDemolitionsTechResearchCost = 15
+kDemolitionsTechResearchTime = 45
+
+kAdvancedArmoryUpgradeCost = 10
+kAdvancedArmoryResearchTime = 60