+++ lua/CompMod/Marine/AdvancedWeapons/Post/TeamInfo.lua
@@ -1,0 +1,33 @@
+TeamInfo.kRelevantTechIdsMarine =
+{
+    kTechId.MinesTech,
+    kTechId.WelderTech,
+    kTechId.GrenadeTech,
+
+    kTechId.AdvancedArmory,
+    kTechId.AdvancedArmoryUpgrade,
+    kTechId.MunitionsTech,
+    kTechId.DemolitionsTech,
+
+    kTechId.Weapons1,
+    kTechId.Weapons2,
+    kTechId.Weapons3,
+    kTechId.Armor1,
+    kTechId.Armor2,
+    kTechId.Armor3,
+
+    kTechId.PrototypeLab,
+    kTechId.JetpackTech,
+    kTechId.ExosuitTech,
+    kTechId.DualMinigunTech,
+
+    kTechId.ARCRoboticsFactory,
+    kTechId.UpgradeRoboticsFactory,
+    kTechId.MACEMPTech,
+    kTechId.MACSpeedTech,
+
+    kTechId.Observatory,
+    kTechId.PhaseTech,
+
+    kTechId.AdvancedMarineSupport,
+}